feat!: add config to user model/type. Add types for contacts. (#111)
- add config to user model -> That introduced a circular dependency that
had to be fixed. Same as before(we are gonna have this everythime types
gets imported into schemas/templates/whatever). Types is meant to be a
leave in the tree but sometimes, like in the reminders stuff you wanna
define an schema based on data, if that data is somehow complex and
import/use types you are screwed out. I cannot think of a definitive
solution except for everything in the same /types or whatever but I
dislike it very much cause it will make harder the code organization.

- added models for phone contact'